pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "coreco"
authors = [{name = "TylorShine", email = "ricka.tylor@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "customtkinter ~=5.1.3",
    "librosa ~=0.10.0.post2",
    "numpy ~=1.24.3",
    "sounddevice ~=0.4.6",
    "soundfile ~=0.12.1",
    "soxr ~=0.3.5",
    "tinydb ~=4.7.1",
    "PyYAML ~=6.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
]
production = [
    "pyinstaller",
]

[project.urls]
Home = "https://github.com/TylorShine/coreco"


[tool.black]
line-length = 100
target-version = ['py310']
include = '\.py$'
skip-string-normalization = true
exclude = '''
(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | venv
    | \.idea
    | \.vscode
    | _build
    | buck-out
    | build
    | dist
    | migrations
)
'''

[tool.isort]
profile = "black"
line_length = 100
